fix: avoid potential data loss in type conversion

+++ crates/cext/src/operators/fermion_operator.rs
@@ -267,7 +267,7 @@ pub unsafe extern "C" fn qf_ferm_op_num_groups(op: *const FermionOperator) -> u3
 pub unsafe extern "C" fn qf_ferm_op_get_groups(
     op: *const FermionOperator,
     groups_out: *mut *mut u32,
-    groups_len: *mut u32,
+    groups_len: *mut u64,
 ) {
     let op = unsafe { const_ptr_as_ref(op) };
     let groups = &op.groups.as_ref().expect(
@@ -308,7 +308,7 @@ pub unsafe extern "C" fn qf_ferm_op_get_groups(
 pub unsafe extern "C" fn qf_ferm_op_set_groups(
     op: *mut FermionOperator,
     groups_in: *const u32,
-    groups_len: u32,
+    groups_len: u64,
 ) {
     let op = unsafe { mut_ptr_as_ref(op) };
     let groups_in = unsafe { const_ptr_as_ref(groups_in) };
